+++ app/(public)/layout.tsx
@@ -1,0 +1,6 @@
+import { BaseLayout } from "@/features/layout/base-layout";
+import type { ReactNode } from "react";
+
+export default function PublicLayout({ children }: { children: ReactNode }) {
+  return <BaseLayout>{children}</BaseLayout>;
+}